Disable external reader project test
